source
git-svn-id: file:///source/coreland/installer/trunk@698 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ fmt.h
@@ -29,6 +29,6 @@ unsigned int fmt_float(char *, float *);
 unsigned int fmt_double(char *, double *);
 
 unsigned long fmt_str(char *, const char *);
-unsigned long fmt_nstr(char *, const char *, unsigned int);
+unsigned long fmt_nstr(char *, const char *, unsigned long);
 
 #endif
